revert table name sanitizing
