--- state/utils/utils_test.go
@@ -1,38 +1,0 @@
-/*
-Copyright 2021 The Dapr Authors
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package utils
-
-import "testing"
-
-func TestValidIdentifier(t *testing.T) {
-	tests := []struct {
-		name string
-		arg  string
-		want bool
-	}{
-		{name: "empty string", arg: "", want: false},
-		{name: "valid characters only", arg: "acz_039_AZS", want: true},
-		{name: "invalid ASCII characters 1", arg: "$", want: false},
-		{name: "invalid ASCII characters 2", arg: "*", want: false},
-		{name: "invalid ASCII characters 3", arg: "hello world", want: false},
-		{name: "non-ASCII characters", arg: "🙃", want: false},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := ValidIdentifier(tt.arg); got != tt.want {
-				t.Errorf("validIdentifier() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
